playing with bitmaps

+++ mosaic/app/src/main/java/com/example/matthew/mosaic/RGB.java
@@ -1,11 +1,11 @@
 package com.example.matthew.mosaic;
 
 public class RGB {
-    int R;
-    int G;
-    int B;
+    float R;
+    float G;
+    float B;
 
-    public RGB(int R_, int G_, int B_){
+    public RGB(float R_, float G_, float B_){
         this.R = R_;
         this.G = G_;
         this.B = B_;
